New changes
All images are in the corect dimensions and everything is in Markdown format. 
The "Built in LED" section has been removed. 
The conclusion has been made more relevant.
I am not satisfied with the quality of the screenshots and will try to increase the quality.
There are now seperate "Circuit" and "Programming the board" sections.