Add changed send trigger to list of changes in IDE 2.x Serial Monitor
In the Arduino IDE 1.x Serial Monitor, sending is triggered via the "Send" button or by pressing the "Enter" key.

There is no "Send" button in the Arduino IDE 2.x Serial Monitor.

The keyboard trigger was changed from "Enter" to "Ctrl + Enter". This is explained to the user by placeholder text in
the input field, but that text disappears as soon as they click on the field, so it is easy to miss that information.
Feedback from users indicates this is the case.

For this reason, it will be important to clearly document the change.