Update link to Library Manager submission instructions
The library submission documentation was moved from the wiki of the Arduino IDE 1.x repository to the dedicated
repository for the Arduino Library Manager registry.

There was still a redirection link at the previously referenced page, but it will be best for this link to send the
reader directly to the relevant information, as is done by this change.