Move IDE 2.x Activity Bar debug button screenshot to appropriate location
Arduino IDE 2.x has two controls related to the integrated debugger:

- The Toolbar button, which is used to initialize the debugger
- The Activity Bar button, which is used to open and close the "Debug" view in the Side Panel

These two things have completely distinct purposes and can not be used interchangeably. Yet they look very similar. Feedback indicates that this complex interface causes confusion for users, and that such confusion will result in the inability to use the debugger, so it is important for the documentation to clearly differentiate between the two.

The debugger tutorial contains separate introductory text and screenshots for each of the buttons. Previously, the content for these two distinct things was ordered like so:

- Introductory text for Activity Bar button
- Introductory text for Toolbar button
- Screenshot for Activity Bar button
- More text related to Toolbar button
- Screenshot for Toolbar button

This order could lead the reader to think the Activity Bar button screenshot was showing the control referenced by the Toolbar button introductory text, since the screenshot came after that introductory text.

The order is hereby changed to be more logical and clear:

- Introductory text for Activity Bar button
- Screenshot for Activity Bar button
- Introductory text for Toolbar button
- More text related to Toolbar button
- Screenshot for Toolbar button