Move IDE 2.x configuration file link above supplemental link
Arduino IDE 2.x allows the user to customize the output of the "Auto Format" feature via a configuration file. This
system is documented in the "Customizing the Auto Formatter Feature" tutorial.

Arduino's standard formatter configuration file is offered as part of the tutorial. Originally, the full YAML code of
the standard configuration was included in the tutorial page itself. Since that consisted as a large block of cryptic
code, it made sense to place it at the bottom of that section of the tutorial, after the section's prose content, following this structure:

- Introduction
- Supplemental note
- Configuration code

Since that time, the code content has been replaced by a link to the source file. After this change, the placement of
this essential link (which is less visibly prominent in the page) makes less sense. The content will be easier to follow
with the new structure:

- Introduction
- Configuration file link
- Supplemental note