Use preferred Arduino IDE 2.x Auto Format menu path
The "Auto Format" feature has always been found under the "Tools" menu in the Arduino IDE. The Arduino IDE 2.x designers
determined it was necessary to move it to the "Edit" menu instead. Although the previous Tools > Auto Format menu path
is still present for now, the plan is to eventually remove it and for Edit > Auto Format to be the sole path to this
feature.

For this reason, the documentation should use the preferred menu path instead of the deprecated path.