Fix Serial Plotter sketches for compatibility with all IDE versions
The sketches provided in the Arduino IDE 2.x Serial Plotter documentation used variable labels that contained spaces.

Support for spaces in labels has been intentionally dropped in order to enable fixes for critical bugs in Serial Plotter
that made it incompatible with the data formats established for years by the Arduino IDE 1.x Serial Plotter. Since that
Serial Plotter never supported spaces in labels, and thus the existing body of sketches did not use them in labels, it
was determined an acceptable tradeoff.

It may be that this enhancement will be implemented at some point in the future, but for now the example sketches in the
docs must use the currently supported data format.