Remove "Board Recovery" section from datasheets of boards without feature
Some Arduino boards have a bootloader that allows them to be recovered from a "soft bricked" state by pressing the reset
button twice.

The bootloader on the boards which do not have a native USB capability don't recognize this double reset. The reason is
because the "soft bricked" state is specific to the native USB boards, so boards without native USB have no need for
this recovery capability.

Previously, bootloader support for double reset was claimed for several boards which do not have such a bootloader.
Since these boards do not have or need any sort of "board recovery" system, the "Board Recovery" section is removed
entirely from their datasheets