Remove outdated info re: Serial Monitor send trigger in IDE 2.x
In previous versions of Arduino IDE 2.x, the keyboard shortcut used to trigger the sending of a message from Serial
Monitor to the board was Ctrl+Enter.

Since this was different from the shortcut used by Arduino IDE 1.x, and user reports indicated this was a significant
source of confusion, a note was added to the Arduino IDE 2.x Serial Monitor tutorial about the change.

The keyboard shortcut was changed to Enter in the Arduino IDE 2.0.1 release, making the tutorial's information about the
changed shortcut obsolete. For this reason, it is hereby removed.

The removed section also documented an additional change to Serial Monitor: the loss of the "Send" button, which is
still the case. The primary purpose for adding the section was to document the keyboard shortcut change and my
impression is that there is no need to mention the removal of the button now that there is an intuitive keyboard
shortcut for sending the message.