Update "J-Link GDB Server" path in IDE 2.x debugger tutorial
Older versions of the "SEGGER J-Link GDB Server" were only available in a 32-bit package for Windows. For this reason,
the path to the server in the tutorial's example debug_custom.json used the `C:\Program Files (x86)` path where the
32-bit package is installed by default on a 64-bit Windows machine.

A 64-bit Windows package has been available starting with the 7.50 release (made on 2021-07-01). This package is
installed under `C:\Program Files` by default (as is the 32-bit package on 32-bit Windows machines).

Most users will have the file at this path. The tutorial will be most friendly to the reader if it uses the most common
path in the example so it is hereby updated to that path.