Use command line version of "J-Link GDB Server" in example configuration file of IDE 2.x debugger tutorial
Installation of the "SEGGER J-Link GDB Server" software is required in order to use a J-Link debug proble with the
Arduino IDE integrated sketch debugger.

This installation includes two variants of the "J-Link GDB Server" tool, each executable with a different filename:

- GUI: `JLinkGDBServer.exe` (Windows) / `JLinkGDBServerExe` (Linux/macOS)
- Command Line: `JLinkGDBServerCL.exe` (Windows) / `JLinkGDBServer` (Linux/macOS)

The user must specify the path to the tool executable in the debug_custom.json configuration file. The Arduino IDE 2.x
debugger tutorial includes a sample debug_custom.json file. Previously this example used the filename of the GUI
variant.

Although technically inappropriate for this application, this variant worked fine with the previous version of the
debugger extension (0.3.10) used by Arduino IDE 2.0.2 and older, and also offered the advantage of having the same
filename as Linux and macOS users will need (confusingly, `JLinkGDBServer` is the filename of the GUI variant on
Windows, and of the command line variant on the other operating systems).

I found that the newer version of the debugger extension (1.5.1) used by Arduino IDE 2.0.3 fails on startup when the GUI
variant is used, while the command line variant works fine. For this reason, the filename used in the example
debug_custom.json file must now be updated to use the command line variant. This should not impact anyone who is
following the tutorial while using an Arduino IDE version 2.0.2 or older because the command line variant works fine
with those versions as well.