Provide filename of "J-Link GDB Server" for each OS in IDE 2.x debugger J-Link tutorial
Installation of the "SEGGER J-Link GDB Server" software is required in order to use a J-Link debug probe with the
Arduino IDE integrated sketch debugger.

The user must specify the path to the command line version of the "J-Link GDB Server" tool executable in their
debug_custom.json configuration files.

The Arduino IDE 2.x debugger tutorial includes a sample debug_custom.json file. This example shows the filename of the
tool on Windows, which would likely be sufficient documentation for the readers using that operating system.

However, this will not be terribly helpful for the readers using other operating systems, for the following reasons:

This installation includes two variants of the "J-Link GDB Server" tool, each executable with a different filename:

- GUI: `JLinkGDBServer.exe` (Windows) / `JLinkGDBServerExe` (Linux/macOS)
- Command Line: `JLinkGDBServerCL.exe` (Windows) / `JLinkGDBServer` (Linux/macOS)

Confusingly, `JLinkGDBServer` is the filename of the GUI variant on Windows, and of the command line variant on the
other operating systems

Although technically inappropriate for this application, either variant worked fine with the previous version of the
debugger extension (0.3.10) used by Arduino IDE 2.0.2 and older. However, I found that the newer version of the debugger
extension (1.5.1) used by Arduino IDE 2.0.3 fails on startup when the GUI variant is used (tested on Windows), while the
command line variant works fine. For this reason, it is now especially important to clearly document the filename to be
used in the configuration file for each operating system.