Add begin() function to example as best practice
The guide instructs the user to place a hardware configuration call, pinMode(), in the class constructor. The problem is that the constructor will execute this in globals scope before Arduino core's main() has time to call the init() function to initialize the hardware. In many cases, this will be fine, but there are cases where the hardware is not ready until after init() is called. To address this, the Morse library class example should include a `begin()` function similar to `Serial.begin()` that is placed in `setup()`. This will encourage students of this guide to follow that best practice.

Note, the Morse.zip file will need to be updated.  I have an updated version here: https://github.com/jasonacox/Morse and will zip and submit that for a replacement if this PR is acceptable.