Correct `arduino-iot-cloud` package name in installation command
Arduino provides a Arduino IoT Cloud client Python package to allow interfacing with Arduino IoT Cloud objects from a
Python script.

The documentation for the package includes a command that can be used to install the package on the host PC. Previously,
the command used an incorrect name for the package, which caused the command to fail:

```
$ pip install arduino-iot-cloud-py
ERROR: Could not find a version that satisfies the requirement arduino-iot-cloud-py (from versions: none)
ERROR: No matching distribution found for arduino-iot-cloud-py
```

The package name in the installation command is hereby corrected.