Update led-matrix.md
The table of function members are clickable, but are not linked to the function members.  The click is actually taking me to the top of the document in some cases.

I have added html anchors to each function member name, and linked them to each corresponding function in the table.