Correct pin numbers for Nano ESP32 LED
Arduino pin numbers are an arbitrary numerical identifier for the I/O pins on the microcontroller. They are mapped to
the lower level identifier in the board's core variant. Generally, Arduino functions that take a pin number argument
require that argument to be an Arduino pin number rather than some lower level identifier for the pin, and that is the
case with the Arduino core functions for the Nano ESP32 board.

Previously, the documentation of the pins connected to the RGB LED on the Nano ESP32 board used the lower level ESP32
GPIO pin numbers in the digitalWrite calls in the code snippet. Since digitalWrite requires the use of an Arduino pin
number, not a GPIO number, that code was wrong and did not control the RGB LED as claimed.

When correcting this error, I chose to use the human friendly constants rather than the bare integer literals (14, 15,
16).