Document special upload procedure in UNO R4 WiFi HID tutorial
When an UNO R4 WiFi board is running a sketch that uses the HID capabilities, the port address changes during the
upload. Since this change does not occur under other conditions, the platform is not configured to handle such a change.
This causes uploads via the standard procedure to fail under these conditions. Since adjusting the configuration to
allow reliable uploads under any conditions would have harmful side effects, the decision was made to leave the
configuration as it is now.

The upload can be accomplished reliably if the user performs a double reset before uploading, but this fact was not
documented in the tutorial about the board's HID capabilities. The missing documentation is hereby added to the
tutorial.