Correct CVD links
Several documentation pages provide links to Arduino's Coordinated Vulnerability Disclosure Policy. This important
document ensures 3rd parties report discovered security vulnerabilities in a responsible manner.

Previously these links targeted a page that was simply a vague overview of the Arduino company's approach to security in
our own work, which is completely irrelevant to a 3rd party wanting to report a vulnerability. The 3rd party would have
had to sift through all that irrelevant information in order to find the tangential link to the actual CVD document.
That unfriendly approach to providing the essential information risks causing some 3rd parties to conclude that Arduino
doesn't have a formal procedure for reporting and resorting to using an insecure or inefficient reporting procedure
instead.

For this reason, it is better to point the links directly to the information they claim to target.