Update Cloud API key creation instructions for new site design
The Arduino Cloud website was recently redesigned. Before the redesign, API keys were managed by selecting "API keys"
from the navigation panel on the Arduino Cloud home page.

That item was removed from the navigation panel and there is now a dedicated page for managing API keys. This means the
instructions based on the old design are no longer correct.

The new design allows us to point the user directly to the relevant page rather than sending them to the home page and
then instructing them in how to navigate to the key management interface. That navigation process is slightly more
complex with the new design so continuing with the previous approach of starting from the home page would require adding
an additional step to the instructions.