Add mandatory prefixes to filenames of downloadables
The system that copies the standard downloadable files to the "resources" folders identifies them by a file suffix which
starts with `-` (e.g., `*-schematics.pdf`). These files were missing the standard `<sku>-` prefix and thus did not match
the pattern and were not copied to the "resources" folders.

The website infrastructure has a system for handling files not covered by the "resources" system. However, unlike the
resource downloadables file copying system of the `arduino/docs-content` repository, the website generator's system
doesn't include the `-` in the file suffix patterns it uses to identify the resource files. This means that if a
downloadable file has a name matching the resource filename pattern, but without the `-`, then it slips through a gap
between the two systems and the download link on the website for these files will be broken.

This is fixed by adding the standard prefix to the filenames of the resource downloadables.

Even though the mismatch between the resource downloadable file identification code in the two systems is a bug, a
differentiating prefix on the filenames is also required to avoid collisions in the resources folder so this change is
beneficial even after the bug in the infrastructure is fixed.