Use placeholder for board name component of "DSpotterSDK_Maker_*" libraries
There is a separate speech recognition library for each of the supported boards. The library name all use the same
"DSpotterSDK_Maker_" prefix, differentiated with a suffix to indicate the target board.

The tutorial contains instructions for opening the library's example sketch. These instructions are intended to be
applicable for use with any of the library variants so it is necessary for the reader to extrapolate the correct library
name. The new approach is to use a placeholder to indicate the board name suffix on the library name.

The common convention of indicating a placeholder via angle brackets was chosen. This causes a visual collision with the
tutorials use of right angle brackets as separators when indicating menu path navigation, as is the standard in Arduino
documentation. In order to avoid this collision, the tutorial is migrated to using the alternative "->" menu path
separator.