Fix broken links to Arduino IDE 2.x docs parent page
The URL of the parent page for the Arduino IDE 2.x documentation was changed from `/software/ide-v2/` to
`software/ide/#ide-v2` through a recent reorganization of the documentation website.

A redirect was set up, but it did not work for these particular links due to them having a trailing slash, which caused
them to lead to the "Oops! There's nothing here" page. The links are hereby updated to point directly to the new URL of
the target page.