2.0.5 (#1758)
* Wip 2.0.5

* test datasheets

* Fix broken links to Arduino IDE 2.x docs parent page

The URL of the parent page for the Arduino IDE 2.x documentation was changed from `/software/ide-v2/` to
`software/ide/#ide-v2` through a recent reorganization of the documentation website.

A redirect was set up, but it did not work for these particular links due to them having a trailing slash, which caused
them to lead to the "Oops! There's nothing here" page. The links are hereby updated to point directly to the new URL of
the target page.

* Content update (Cloud Editor screenshot update) (#1755)

* Updating trademarks for PMC

Updating the trademark symbols for the WiFi and Bluetooth trademark in the hardware page of PMC

* updated store link to plant watering kit

* Update content/hardware/09.kits/maker/plant-watering-kit/tutorials/iot-smart-garden-setup-with-nanorp2040/content.md

Co-authored-by: Josefine Hansson <66409231+jhansson-ard@users.noreply.github.com>

* 2.0.5

* Remove schematics tests

---------

Co-authored-by: per1234 <accounts@perglass.com>
Co-authored-by: José Antonio Bagur Nájera <josea.bagur@gmail.com>
Co-authored-by: Josefine Hansson <66409231+jhansson-ard@users.noreply.github.com>
Co-authored-by: Josefine Hansson <j.hansson@arduino.cc>
Co-authored-by: Jacob Hylén <60390259+jacobhylen@users.noreply.github.com>