Add instructions for selecting debug probe from "Programmer" menu in IDE 2.x debugger tutorials
A major reworking of the debugger configuration system was released in Arduino IDE 2.3.0. One of the user facing changes
is the ability to configure the IDE for the debug probe hardware in use via the "Tools > Programmer" menu selection.

As a side effect of that change, it is now mandatory for the user to make the appropriate selection from that menu in
order to use the debugger. If no selection is made, the IDE disables the "Start Debugging" button. For this reason, it
will be important for the debugger tutorials to include this.

Arduino IDE versions prior to 2.3.0 did not use the "Programmer" selection for configuration of the debugger. Although
selecting a programmer before starting the debugger is not mandatory for users of older versions of the IDE, it also
will not cause any harm and that menu was present in those versions of the IDE as well. So the inclusion of a
"Programmer" menu selection step in the tutorials does not break their compatibility with previous versions of Arduino
IDE 2.x.