Update IDE 2.x J-Link tutorial for new debug_custom.json data format
A major reworking of the Arduino IDE 2.x debugger configuration system was released in Arduino IDE 2.3.0. This included
a breaking change to the data structure of the debug_custom.json file that allows the user to adjust the debugger
configuration.

The tutorial for using the Arduino IDE 2.x debugger with a J-Link debug probe requires the use of the debug_custom.json
feature. The change to the file's data structure rendered that part of the tutorial outdated. It is hereby updated to
use the new data structure.

Including instructions for the different debug_custom.json structure that would be required for users of versions of
Arduino IDE 2.x older than 2.3.0 would have added significant complexity to an already dense tutorial, making it
significantly less approachable to the reader and prone to misinterpretation. For this reason, the decision was made to
specify a requirement of Arduino IDE 2.3.0 as the minimum for the tutorial.