Add alternative fields to J-Link debug_custom.json to workaround tutorial breaking IDE bug
The tutorial for using the Arduino IDE 2.x debugger with a J-Link debug probe requires the use of a debug_custom.json
file to adjust the debugger configuration in order to allow it to be used with the J-Link.

A regression was introduced in Arduino IDE 2.3.0 that causes the configuration defined in certain fields of
debug_custom.json to no longer have an effect, causing the tutorial to no longer be usable.

A fix for the bug is now in progress, but it will likely be some time before this can be shipped in a new Arduino IDE
release. For this reason, it will be worth adjusting the contents of the debug_custom.json file provided in the tutorial
to work around the bug, allowing it to work with the current versions of Arduino IDE.

Once a new release of Arduino IDE is made with the bug fixed, this commit should be reverted and the minimum IDE version
specified in the tutorial bumped. However the workaround is applied in a manner that will not cause any harm even when
present in the debug_custom.json file while a fixed version of Arduino IDE is in use.