Move information re: SoftwareSerial pins to table
Previously, the information about which pins could be used with the SoftwareSerial library was documented in prose form.

Although reasonable in the context of only needing to cover a few boards as was the case at the time the documentation
was established, this approach does not scale as it is necessary to add information for more boards over time. For this
reason, the information is migrated to a table in preparation for addition of the missing documentation for the newer
boards.