Add all supported boards to SoftwareSerial pins table
Since the time the SoftwareSerial library documentation was written, Arduino has created several new products supported
by the library. The information about which pins can be used on these board must be documented.

After this, the only currently active product with SoftwareSerial library support not included in the table is the
Portenta C33. From the information in the library's "SerialEcho" example, it might be inferred that the same limitations
as the UNO R4 boards apply to this board as well, but I was not able to verify this due to not having the hardware on
hand.