Replace links to ephemeral file URLs with persistent "resource" links
In addition to the Markdown content source files, this repository hosts non-text files. These are linked to in various
places in the content.

The website publishing system hosts these non-text files in paths under the `static` subfolder of the website. The path
of the published file includes the file's MD5 hash. This means that the URL will change every time the file is modified.
For this reason, hardcoding links to these URLs in the content should be avoided whenever possible.

The repository contains a special `static` folder. The files under that folder are published under the root of the
website, using the same paths as are present in the `static` folder at the time of publishing. In addition to the files
that are hosted in the repository's `static` folder, the repository infrastructure also finds all files under the
`content/hardware` folder with suffix matching standardized patterns used by the common product "resource" files and
copies them to the appropriate locations under the `static` folder before publishing the website. Links should always
target the files at this location when available instead of using the ephemeral URLS.