Replace links to ephemeral file URLs with relative paths
In addition to the Markdown content source files, this repository hosts non-text files. These are linked to in the
content.

The website publishing system hosts these non-text files in paths under the `static` subfolder of the website. The path
of the published file includes the file's MD5 hash. This means that the URL will change every time the file is modified.
For this reason, hardcoding links to these URLs in the content should be avoided whenever possible.

The website generation system automatically replaces links to relative paths with the equivalent ephemeral URL under the
`static` subfolder of the website. This means that a link to the relative path of the file is functionally equivalent to
linking the hardcoded ephemeral URL.

Although a relative path link is still prone to breakage if the content or target file is moved to a different location
in the repository, this will probably occur less frequently than modifications to the files, making the relative path
approach superior to the alternative of hardcoding the ephemeral URL.