Use standardized pin macros in Nano ESP32 "default" sketch
This sketch is intended to produce the LED blinking behavior the board has when shipped. Previously, bespoke pin name
macros were defined in the sketch, which were problematic:

The green and blue LED pin numbers were only appropriate for the first limited run of the board that used a different
RGB LED than the LED on subsequent runs of the board, which the majority of customers will have.

The pin numbers were not correct when the development tool had the "By Arduino pin (default)" pin numbering
configuration.

The "Arduino ESP32 Boards" core provides pin macros for all four of the LEDs. Using these macros in the sketch sets an
example of best practices programming for the reader and fixes both the problems that were present in the previous
version of the sketch:

The LED_GREEN and LED_BLUE are the correct pin numbers for the LEDs on the majority of the Nano ESP32 boards.

These macros are configured to be correct for both pin numbering configurations.