content.md: disable RCP auto-update from sdkconfig.defaults
While following this documentation today, I noticed that it didn't work. ESP32 board complained that it's trying to update RCP and failing to do so.

I noticed that the sdkconfig line for ```CONFIG_OPENTHREAD_BR_AUTO_UPDATE_RCP``` is automatically uncommented / overwritten when running ```idf.py build```. That config is enabled inside sdconfig.defaults file and it seems that build step writes atleast some defaults to sdkconfig.

This commit adds additional step for commenting out that RCP auto update setting from sdkconfig.defaults file as well, which fixed this issue for me.

For context I was following this guide on 9th of August 2025 with POP!_OS 22.04 LTS 64-bit. Fresh git pulls for every step, just as mentioned in the guide.