docs: Add missing read() method documentation to I2S library
Fixes #2583

Added comprehensive documentation for the read() method in the I2S library.

The read() method was being used in examples but was not documented in the API reference. This commit adds:
- Method description
- Two syntax variants (single sample and buffer read)
- Parameter documentation
- Return value documentation

This resolves the issue where users could not find documentation for the read() method despite it being used in code examples.